Apply migration changes and ingredient became item.

+++ database/migrations/2025_02_17_222757_add_category_id_to_recipes_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('recipes', function (Blueprint $table) {
-            $table->foreignId('category_id')->constrained()->onDelete('cascade');
+            if (!Schema::hasColumn('recipes', 'category_id')) {
+                $table->foreignId('category_id')->constrained()->onDelete('cascade');
+            }
         });  //'category_id' は 'categoriesテーブル' の 'id' を参照する外部キーです
     }
